+++ packages/vsce/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cics-extension-for-zowe" extension will be documented in this file.
 
+## Recent Changes
+
+- Feature: Add React-based attributes view in the panel to replace existing attributes editor [#281](https://github.com/zowe/cics-for-zowe-client/issues/281)
+
 ## `3.7.0`
 
 - BugFix: Allow credentials to be used by the current session only with `autoStore: false` [#224](https://github.com/zowe/cics-for-zowe-client/issues/224)